deleteEventRepo func throws error when person other than orginal creator is trying to deleteEvet and message for validationerror in deletePostRepo improved src\repositories\post.repository.ts

+++ src/repositories/post.repository.ts
@@ -13,9 +13,9 @@ export const deletePostRepo = async (user_id: string, post_id: string): Promise<
 	const client = await pool.connect();
 	try {
 		await client.query("BEGIN;");
-		const { rows } = await client.query("SELECT FROM posts WHERE user_id = $1 AND id = $2", [user_id, post_id]);
+		const { rows } = await client.query("SELECT id, (SELECT COUNT (id) FROM posts WHERE id = $2) as count FROM posts WHERE user_id = $1 AND id = $2 LIMIT 1", [user_id, post_id]);
 		if (!rows[0]) {
-			throw new ValidationError("Unauthorized person is trying to delete post!", 401);
+			throw new ValidationError("Unauthorized post operation!", 401);
 		}
 		await client.query("DELETE FROM likes WHERE post_id = $1", [post_id]);
 		await client.query("DELETE FROM comments WHERE post_id = $1", [post_id]);
@@ -35,8 +35,6 @@ export const loadMyPostsRepo = async (user_id: string, offset: number): Promise<
 		"WITH postsArr AS (SELECT posts.id, posts.group_id, posts.title, posts.text, posts.picture, posts.attachment, posts.created_at, posts.type, (SELECT COUNT(user_id) FROM likes WHERE post_id = posts.id) AS likes, (SELECT COUNT (user_id) FROM likes WHERE user_id = $1 and post_id = posts.id) as likedByUser, (SELECT COUNT (id) FROM comments WHERE post_id = posts.id) AS comments, (SELECT username FROM users WHERE id = comments.user_id) AS commentusername, comments.id AS commentid, comments.text AS commenttext, comments.picture AS commentpicture, comments.attachment AS commentattachment, comments.created_at AS commentcreated_at, ROW_NUMBER() OVER(PARTITION BY posts.id ORDER BY posts.created_at DESC) as row_num FROM posts  FULL JOIN comments ON posts.id = comments.post_id WHERE posts.user_id = $1 ORDER BY posts.created_at DESC) SELECT id, group_id, title, text, picture, attachment, created_at, type, likes, likedByUser, comments, commentusername, commentid, commenttext, commentpicture, commentattachment, commentcreated_at FROM postsArr WHERE row_num = 1 LIMIT 10 OFFSET $2",
 		[user_id, offset],
 	);
-	console.log(rows);
-
 	return rows;
 };
 
@@ -108,6 +106,10 @@ export const deleteEventRepo = async (user_id: string, event_id: string): Promis
 	const client = await pool.connect();
 	try {
 		await client.query("BEGIN;");
+		const { rows } = await client.query("SELECT id FROM posts WHERE user_id = $1 AND id = $2 LIMIT 1", [user_id, event_id]);
+		if (!rows[0]) {
+			throw new ValidationError("Unauthorized post operation!", 401);
+		}
 		await client.query("DELETE FROM users_events WHERE event_id = $1;", [event_id]);
 		await client.query("DELETE FROM likes WHERE post_id = $1;", [event_id]);
 		await client.query("DELETE FROM comments WHERE post_id = $1;", [event_id]);
